+++ mds/provider_poller/poller/v0_2.py
@@ -2,6 +2,8 @@
 Pulling data for registered providers
 
 This is the opposite of provider pushing their data to the agency API.
+
+TODO(hcauwelier) Should be deleted ASAP
 """
 import datetime
 import enum
@@ -23,8 +25,8 @@ from mds import enums
 from mds import models
 from mds import utils
 from mds.provider_mapping import PROVIDER_REASON_TO_AGENCY_EVENT
-from .oauth2_store import OAuth2Store
-from .translation import translate_v0_2_to_v0_4
+from ..oauth2_store import OAuth2Store
+from .translation import translate_v0_2
 
 
 MDS_CONTENT_TYPE = "application/vnd.mds.provider+json"
@@ -57,8 +59,11 @@ class StatusChangesPoller:
         to_cursor: timestamp or int, upper limit
     """
 
-    def __init__(self, provider, cursor=None, from_cursor=None, to_cursor=None):
+    def __init__(
+        self, provider, api_version, cursor=None, from_cursor=None, to_cursor=None
+    ):
         self.provider = provider
+        self.api_version = api_version
         self.cursor = cursor
         self.from_cursor = from_cursor
         self.to_cursor = to_cursor
@@ -74,21 +79,14 @@ class StatusChangesPoller:
             logger.debug("Provider %s has no URL, skipping.", self.provider.name)
             return
 
-        # We request a specific version of the Provider API
-        # TODO(hcauwelier) make it mandatory, see SMP-1673
-        api_version_raw = self.provider.api_configuration.get(
-            "api_version", enums.DEFAULT_PROVIDER_API_VERSION
+        logger.debug(
+            "Start polling provider %s using version %s",
+            self.provider.name,
+            self.api_version,
         )
-        # We store the enum key, which cannot be a numeric identifier
-        # It doubles as a validity check
-        api_version = enums.MDS_VERSIONS[api_version_raw].value
+        self._poll_status_changes()
 
-        logger.debug("Polling %s using version %s", self.provider.name, api_version)
-
-        getattr(self, "_poll_status_changes_%s" % api_version_raw)(api_version)
-
-    # TODO(hcauwelier) Should be deleted ASAP
-    def _poll_status_changes_v0_2(self, api_version):
+    def _poll_status_changes(self):
         # Start where we left, it's all based on providers sorting by start_time
         # (but we would miss telemetries older than start_time saved after we polled).
         # For those that support it, use the recorded time field or equivalent.
@@ -164,9 +162,9 @@ class StatusChangesPoller:
 
         # Pagination
         while next_url:
-            body = self._get_body(next_url, api_version)
+            body = self._get_body(next_url)
             # Translate older versions of data
-            translated_data = translate_v0_2_to_v0_4(body["data"])
+            translated_data = translate_v0_2(body["data"])
             status_changes = translated_data["status_changes"]
             if not status_changes:
                 break
@@ -217,242 +215,8 @@ class StatusChangesPoller:
                     + f"\tLast skip: {skip_polled}."
                 )
 
-    def _poll_status_changes_v0_3(self, api_version):
-        # Start where we left, it's all based on providers sorting by start_time
-        # (but we would miss telemetries older than start_time saved after we polled).
-        # For those that support it, use the recorded time field or equivalent.
-        polling_cursor = self.cursor or self.provider.api_configuration.get(
-            "polling_cursor", POLLING_CURSORS.start_time.name
-        )
-        logger.info(
-            f"Starting polling {self.provider} using the field: {polling_cursor}.\n"
-            "Current state:\n"
-            + (
-                (
-                    f"\tLast event_time: {str(self.provider.last_event_time_polled)},\n"
-                    + f"\tLast recorded: {str(self.provider.last_recorded_polled)}\n"
-                    + f"\tLast skip: {self.provider.last_skip_polled}."
-                )
-                if not self.cursor
-                else f"\tCursor value: {self.from_cursor}."
-            )
-        )
-
-        params = {}
-
-        # Cursor param (raise if typo)
-        param_name = POLLING_CURSORS[polling_cursor].value
-
-        if polling_cursor == POLLING_CURSORS.total_events.name:
-            # Resume from the last line fetched
-            params[param_name] = self.from_cursor or self.provider.last_skip_polled
-        elif polling_cursor == POLLING_CURSORS.start_recorded.name:
-            # Resume from the last "recorded" value
-            params[param_name] = utils.to_mds_timestamp(
-                self.from_cursor or self.provider.last_recorded_polled
-            )
-        elif polling_cursor == POLLING_CURSORS.start_time.name:
-            # Resume from the last "event_time" value
-            last_event_time_polled = self.provider.last_event_time_polled
-            if not last_event_time_polled:
-                last_event_time_polled = timezone.now() - datetime.timedelta(
-                    getattr(settings, "PROVIDER_POLLER_LIMIT", 90)
-                )
-
-            # But we now apply a "lag" before actually polling,
-            # leaving time for the provider to collect data from its devices
-            polling_lag = self.provider.api_configuration.get("provider_polling_lag")
-            if polling_lag:
-                polling_lag = parse_duration(polling_lag)
-                if (timezone.now() - last_event_time_polled) < polling_lag:
-                    logger.debug("Still under the polling lag, back to sleep.")
-                    return
-
-            params[param_name] = utils.to_mds_timestamp(
-                self.from_cursor or last_event_time_polled
-            )
-
-        # Provider-specific params to optimise polling
-        try:
-            params.update(self.provider.api_configuration["status_changes_params"])
-        except KeyError:
-            pass
-
-        next_url = urllib.parse.urljoin(self.provider.base_api_url, "status_changes")
-        if self.provider.api_configuration.get("trailing_slash"):
-            next_url += "/"
-
-        if params:
-            next_url = "%s?%s" % (next_url, urllib.parse.urlencode(params))
-
-        skip_polled = (
-            self.from_cursor
-            if self.from_cursor and self.cursor == POLLING_CURSORS.total_events.name
-            else None
-        )
-
-        # Pagination
-        while next_url:
-            body = self._get_body(next_url, api_version)
-            # MDS 0.3 is backwards compatible with 0.4
-            status_changes = body["data"]["status_changes"]
-            if not status_changes:
-                break
-
-            next_url = body.get("links", {}).get("next")
-
-            # A transaction for each "page" of data
-            with transaction.atomic():
-                # We get the maximum of the recorded and event_types
-                # from the status changes
-                event_time_polled, recorded_polled = self._process_status_changes(
-                    status_changes
-                )
-
-                if self.cursor:
-                    if self.cursor == POLLING_CURSORS.total_events.name:
-                        skip_polled += len(status_changes)
-                        if skip_polled >= self.to_cursor:
-                            break
-                    elif self.cursor == POLLING_CURSORS.start_recorded.name:
-                        if recorded_polled >= self.to_cursor:
-                            break
-                    elif self.cursor == POLLING_CURSORS.start_time.name:
-                        if event_time_polled >= self.to_cursor:
-                            break
-                else:
-                    # We get the new skip from the number of status changes
-                    skip_polled = (
-                        self.provider.last_skip_polled + len(status_changes)
-                        if self.provider.last_skip_polled
-                        else len(status_changes)
-                    )
-                    self.provider.last_event_time_polled = event_time_polled
-                    self.provider.last_recorded_polled = recorded_polled
-                    self.provider.last_skip_polled = skip_polled
-                    self.provider.save(
-                        update_fields=[
-                            "last_event_time_polled",
-                            "last_recorded_polled",
-                            "last_skip_polled",
-                        ]
-                    )
-
-                logger.info(
-                    f"Polled page using cursor: {polling_cursor}. New state:\n"
-                    + f"\tLast event_time: {str(event_time_polled)},\n"
-                    + f"\tLast recorded: {str(recorded_polled)}\n"
-                    + f"\tLast skip: {skip_polled}."
-                )
-
-    def _poll_status_changes_v0_4(self, api_version):
-        # Start where we left, it's all based on providers sorting by start_time
-        # (but we would miss telemetries older than start_time saved after we polled).
-        # For those that support it, use the recorded time field or equivalent.
-        polling_cursor = self.cursor or self.provider.api_configuration.get(
-            "polling_cursor", POLLING_CURSORS.start_time.name
-        )
-        if polling_cursor != POLLING_CURSORS.start_time.name:
-            raise ValueError('Only "start_time" cursor is supported in MDS 0.4+')
-        logger.info(
-            f"Starting polling {self.provider} using the field: {polling_cursor}.\n"
-            "Current state:\n"
-            + (
-                (f"\tLast event_time: {str(self.provider.last_event_time_polled)}.")
-                if not self.cursor
-                else f"\tCursor value: {self.from_cursor}."
-            )
-        )
-
-        params = {}
-
-        # Resume from the last "event_time" value
-        last_event_time_polled = self.provider.last_event_time_polled
-        if not last_event_time_polled:
-            last_event_time_polled = timezone.now() - datetime.timedelta(
-                getattr(settings, "PROVIDER_POLLER_LIMIT", 90)
-            )
-
-        # But we now apply a "lag" before actually polling,
-        # leaving time for the provider to collect data from its devices
-        polling_lag = self.provider.api_configuration.get("provider_polling_lag")
-        if polling_lag:
-            polling_lag = parse_duration(polling_lag)
-            if (timezone.now() - last_event_time_polled) < polling_lag:
-                logger.info("Still under the polling lag, back to sleep.")
-                return
-
-        # The MDS 0.4 Provider API got two endpoints:
-        # - /events for the real-time or so data (formerly /status_changes)
-        #   but limited to two weeks of history
-        # - /status_changes for all the data except the current hour
-        # If we're catching up far in time, begin by polling /status_changes
-        realtime_threshold = parse_duration(  # Default is 9 days
-            self.provider.api_configuration.get("realtime_threshold", "P9D")
-        )
-        next_event_time = last_event_time_polled + datetime.timedelta(hours=1)
-        if (timezone.now() - next_event_time) > realtime_threshold:
-            # We have to query the archived status changes with another format
-            logger.info("last_event_time_polled is too old, asking archives")
-            # We're done with the events of the last hour, ask the next hour
-            params["event_time"] = next_event_time.isoformat()[: len("YYYY-MM-DDTHH")]
-        else:
-            # Query the real-time endpoint as usual
-            params["start_time"] = utils.to_mds_timestamp(
-                self.from_cursor or last_event_time_polled
-            )
-
-        # Provider-specific params to optimise polling
-        try:
-            params.update(self.provider.api_configuration["status_changes_params"])
-        except KeyError:
-            pass
-
-        endpoint = "events"  # The new name for the real-time events endpoint
-        if "event_time" in params:
-            # We asked the archived status changes instead
-            endpoint = "status_changes"
-        next_url = urllib.parse.urljoin(self.provider.base_api_url, endpoint)
-        if self.provider.api_configuration.get("trailing_slash"):
-            next_url += "/"
-
-        if params:
-            next_url = "%s?%s" % (next_url, urllib.parse.urlencode(params))
-
-        # Pagination
-        while next_url:
-            body = self._get_body(next_url, api_version)
-            # No translation needed as long as 0.4 is the latest version
-            status_changes = body["data"]["status_changes"]
-            next_url = body.get("links", {}).get("next")
-
-            # A transaction for each "page" of data
-            with transaction.atomic():
-                if status_changes:
-                    # We get the maximum values from the status changes
-                    event_time_polled, _ = self._process_status_changes(status_changes)
-                elif endpoint == "status_changes":
-                    # This hour frame of archives didn't contain results
-                    event_time_polled = next_event_time
-                else:
-                    # Try again from this point later
-                    break
-
-                if self.cursor:
-                    if self.cursor == POLLING_CURSORS.start_time.name:
-                        if event_time_polled >= self.to_cursor:
-                            break
-                else:
-                    self.provider.last_event_time_polled = event_time_polled
-                    self.provider.save(update_fields=["last_event_time_polled"])
-
-                logger.info(
-                    f"Polled page using cursor: {polling_cursor}. New state:\n"
-                    + f"\tLast event_time: {str(event_time_polled)}."
-                )
-
     @retry(stop_max_attempt_number=2)
-    def _get_body(self, url, api_version):
+    def _get_body(self, url):
         authentication_type = self.provider.api_authentication.get("type")
         if authentication_type in (None, "", "none"):
             client = requests
@@ -463,18 +227,12 @@ class StatusChangesPoller:
 
         headers = {}
 
-        # TODO(hcauwelier) just fork the code for MDS 0.4
-        if api_version == "0.4":
-            # We only accept the version we expect from that provider
-            # And it should reject it when it bumps to the new version
-            headers["Accept"] = "%s;version=%s" % (MDS_CONTENT_TYPE, api_version)
-        else:
-            # TODO(hcauwelier) some server implementations are flawed
-            # leave the standard content type for now
-            headers["Accept"] = "application/json,%s;version=%s" % (
-                MDS_CONTENT_TYPE,
-                api_version,
-            )
+        # MDS 0.2 didn't have versioning headers,
+        # Use the standard content type first
+        headers["Accept"] = "application/json,%s;version=%s" % (
+            MDS_CONTENT_TYPE,
+            self.api_version,
+        )
 
         logger.debug("Polling provider on URL %s with headers %s", url, headers)
         response = client.get(url, timeout=30, headers=headers)
@@ -497,11 +255,11 @@ class StatusChangesPoller:
             assert content_type == MDS_CONTENT_TYPE
             params = dict(param.split("=") for param in params)
             received = params.get("version", "unspecified")
-        if received != api_version:
+        if received != self.api_version:
             logger.warning(
                 "Provider %s didn't respond in version %s but %s",
                 self.provider.name,
-                api_version,
+                self.api_version,
                 received,
             )
