+++ tests/management/commands/test_poll_providers_v0_4.py
@@ -10,17 +10,19 @@ from django.utils import timezone
 from mds import enums
 from mds import factories
 from mds import models
-from mds.provider_mapping import (
-    PROVIDER_REASON_TO_AGENCY_EVENT,
-    PROVIDER_EVENT_TYPE_REASON_TO_EVENT_TYPE,
-)
+
+from . import utils
+
+# Note: testing with the default "POLLER_CREATE_REGISTER_EVENTS = False"
 
 
 @pytest.mark.django_db
-def test_poll_provider_batch(client, settings, requests_mock):
+def test_poll_provider_batch(client, requests_mock):
     """A single provider with two pages of status changes."""
-    settings.POLLER_CREATE_REGISTER_EVENTS = True
-    provider = factories.Provider(base_api_url="http://provider")
+    provider = factories.Provider(
+        base_api_url="http://provider",
+        api_configuration__api_version=enums.MDS_VERSIONS.v0_4.name,
+    )
     # The first device received already exists
     device1 = factories.Device(provider=provider)
     expected_event1 = factories.EventRecord.build(
@@ -34,57 +36,62 @@ def test_poll_provider_batch(client, settings, requests_mock):
     )
     stdout, stderr = io.StringIO(), io.StringIO()
 
+    # As we're starting before the threshold, we'll poll the archives endpoint
     url = urllib.parse.urljoin(provider.base_api_url, "/status_changes")
     next_page = "%s?page=2" % url
     requests_mock.get(
         url,
-        json=make_response(
+        json=utils.make_response(
             provider,
             device1,
             expected_event1,
             event_type_reason="service_start",
             associated_trip="e7a9d3aa-68ea-4666-8adf-7bad40e49805",
             next_page=next_page,
+            version="0.4.0",
         ),
     )
     requests_mock.get(
         next_page,
-        json=make_response(
+        json=utils.make_response(
             provider,
             expected_device2,
             expected_event2,
             event_type_reason="maintenance_pick_up",
+            version="0.4.0",
         ),
     )
     call_command("poll_providers", "--raise-on-error", stdout=stdout, stderr=stderr)
 
-    assert_command_success(stdout, stderr)
+    utils.assert_command_success(stdout, stderr)
 
     event1 = device1.event_records.get()
-    assert_event_equal(event1, expected_event1)
+    utils.assert_event_equal(event1, expected_event1)
 
     # The second device was created on the fly
     device2 = models.Device.objects.get(pk=expected_device2.pk)
     assert device2.saved_at is not None
-    # With a fake register event and the actual event
-    event2_register, event2_regular = device2.event_records.order_by("timestamp")
+    event2 = device2.event_records.get()
 
-    assert event2_register.event_type == enums.EVENT_TYPE.register.name
-    assert event2_register.properties == {"created_on_register": True}
-    assert_event_equal(event2_regular, expected_event2)
-    assert_device_equal(device2, expected_device2)
+    utils.assert_event_equal(event2, expected_event2)
+    utils.assert_device_equal(device2, expected_device2)
 
 
 @pytest.mark.django_db
-def test_several_providers(client, django_assert_num_queries, settings, requests_mock):
+def test_several_providers(client, django_assert_num_queries, requests_mock):
     """Two providers this time."""
-    settings.POLLER_CREATE_REGISTER_EVENTS = True
-    provider1 = factories.Provider(base_api_url="http://provider1")
+    provider1 = factories.Provider(
+        base_api_url="http://provider1",
+        api_configuration__api_version=enums.MDS_VERSIONS.v0_4.name,
+    )
     device1 = factories.Device.build(provider=provider1)
     expected_event1 = factories.EventRecord.build(
         event_type=enums.EVENT_TYPE.provider_drop_off.name
     )
-    provider2 = factories.Provider(base_api_url="http://provider2")
+    provider2 = factories.Provider(
+        base_api_url="http://provider2",
+        api_configuration__api_version=enums.MDS_VERSIONS.v0_4.name,
+    )
     device2 = factories.Device.build(provider=provider2)
     expected_event2 = factories.EventRecord.build(
         event_type=enums.EVENT_TYPE.trip_start.name
@@ -97,46 +104,48 @@ def test_several_providers(client, django_assert_num_queries, settings, requests
     n += (
         2  # Savepoint/release for each provider
         + 1  # Insert missing devices
-        + 1  # Insert fake register event
         + 1  # Insert missing event records
         + 1  # Update last start time polled
     ) * 2  # For each provider
     with django_assert_num_queries(n):
         requests_mock.get(
             urllib.parse.urljoin(provider1.base_api_url, "/status_changes"),
-            json=make_response(
+            json=utils.make_response(
                 provider1,
                 device1,
                 expected_event1,
                 event_type_reason="rebalance_drop_off",
+                version="0.4.0",
             ),
         )
         requests_mock.get(
             urllib.parse.urljoin(provider2.base_api_url, "/status_changes"),
-            json=make_response(
-                provider2, device2, expected_event2, event_type_reason="maintenance"
+            json=utils.make_response(
+                provider2,
+                device2,
+                expected_event2,
+                event_type_reason="maintenance",
+                version="0.4.0",
             ),
         )
         call_command("poll_providers", "--raise-on-error", stdout=stdout, stderr=stderr)
 
-    assert_command_success(stdout, stderr)
+    utils.assert_command_success(stdout, stderr)
 
-    event1_register, event1_regular = device1.event_records.order_by("timestamp")
-    assert event1_register.event_type == enums.EVENT_TYPE.register.name
-    assert event1_register.properties == {"created_on_register": True}
-    assert_event_equal(event1_regular, expected_event1)
+    event1 = device1.event_records.get()
+    utils.assert_event_equal(event1, expected_event1)
 
-    event2_register, event2_regular = device2.event_records.order_by("timestamp")
-    assert event2_register.event_type == enums.EVENT_TYPE.register.name
-    assert event2_register.properties == {"created_on_register": True}
-    assert_event_equal(event2_regular, expected_event2)
+    event2 = device2.event_records.get()
+    utils.assert_event_equal(event2, expected_event2)
 
 
 @pytest.mark.django_db
-def test_follow_up(client, settings, requests_mock):
+def test_follow_up(client, requests_mock):
     """Catching up new telemetries from the last one we got."""
-    settings.POLLER_CREATE_REGISTER_EVENTS = True
-    event = factories.EventRecord(device__provider__base_api_url="http://provider")
+    event = factories.EventRecord(
+        device__provider__base_api_url="http://provider",
+        device__provider__api_configuration__api_version=enums.MDS_VERSIONS.v0_4.name,
+    )
     device = event.device
     provider = device.provider
     provider.last_event_time_polled = event.timestamp
@@ -148,32 +157,36 @@ def test_follow_up(client, settings, requests_mock):
 
     # Mocking must fail if the command does not make the expected query
     requests_mock.get(
-        urllib.parse.urljoin(provider.base_api_url, "/status_changes"), status_code=400,
+        urllib.parse.urljoin(provider.base_api_url, "/events"), status_code=400,
     )
     requests_mock.get(
         urllib.parse.urljoin(
             provider.base_api_url,
-            "/status_changes?%s"
+            "/events?%s"
             % urllib.parse.urlencode(
                 {"start_time": round(event.timestamp.timestamp() * 1000)}
             ),
         ),
-        json=make_response(
-            provider, device, expected_event, event_type_reason="service_end"
+        json=utils.make_response(
+            provider,
+            device,
+            expected_event,
+            event_type_reason="service_end",
+            version="0.4.0",
         ),
     )
     call_command("poll_providers", "--raise-on-error", stdout=stdout, stderr=stderr)
 
-    assert_command_success(stdout, stderr)
+    utils.assert_command_success(stdout, stderr)
 
+    # The existing one and the new one
     assert device.event_records.count() == 2
     event = device.event_records.latest("timestamp")
-    assert_event_equal(event, expected_event)
+    utils.assert_event_equal(event, expected_event)
 
 
 @pytest.mark.django_db
-def test_poll_provider_v0_4_archives(client, requests_mock):
-    # Note: testing with the default "POLLER_CREATE_REGISTER_EVENTS = False"
+def test_poll_provider_archives(client, requests_mock):
     provider = factories.Provider(
         base_api_url="http://provider",
         api_configuration__api_version=enums.MDS_VERSIONS.v0_4.name,
@@ -189,7 +202,7 @@ def test_poll_provider_v0_4_archives(client, requests_mock):
     status_changes = urllib.parse.urljoin(provider.base_api_url, "/status_changes")
     requests_mock.get(
         status_changes,
-        json=make_response(
+        json=utils.make_response(
             provider,
             expected_device,
             expected_event,
@@ -203,7 +216,7 @@ def test_poll_provider_v0_4_archives(client, requests_mock):
     )
     call_command("poll_providers", "--raise-on-error", stdout=stdout, stderr=stderr)
 
-    assert_command_success(stdout, stderr)
+    utils.assert_command_success(stdout, stderr)
     # The last poller cursor is updated
     provider = models.Provider.objects.get(pk=provider.pk)
     assert provider.last_event_time_polled is not None
@@ -213,13 +226,12 @@ def test_poll_provider_v0_4_archives(client, requests_mock):
     # The actual event
     event = device.event_records.get()
     assert event.event_type == enums.EVENT_TYPE.service_start.name
-    assert_event_equal(event, expected_event)
-    assert_device_equal(device, expected_device)
+    utils.assert_event_equal(event, expected_event)
+    utils.assert_device_equal(device, expected_device)
 
 
 @pytest.mark.django_db
-def test_poll_provider_v0_4_archives_resume(client, requests_mock):
-    # Note: testing with the default "POLLER_CREATE_REGISTER_EVENTS = False"
+def test_poll_provider_archives_resume(client, requests_mock):
     last_event_time_polled = datetime.datetime(
         2019, 10, 16, 14, 40, 35, tzinfo=timezone.utc
     )
@@ -239,7 +251,7 @@ def test_poll_provider_v0_4_archives_resume(client, requests_mock):
     status_changes = urllib.parse.urljoin(provider.base_api_url, "/status_changes")
     requests_mock.get(
         status_changes,
-        json=make_response(
+        json=utils.make_response(
             provider,
             expected_device,
             expected_event,
@@ -253,7 +265,7 @@ def test_poll_provider_v0_4_archives_resume(client, requests_mock):
     )
     call_command("poll_providers", "--raise-on-error", stdout=stdout, stderr=stderr)
 
-    assert_command_success(stdout, stderr)
+    utils.assert_command_success(stdout, stderr)
     # The last poller cursor is updated
     provider = models.Provider.objects.get(pk=provider.pk)
     assert provider.last_event_time_polled > last_event_time_polled
@@ -263,13 +275,12 @@ def test_poll_provider_v0_4_archives_resume(client, requests_mock):
     # The actual event
     event = device.event_records.get()
     assert event.event_type == enums.EVENT_TYPE.service_start.name
-    assert_event_equal(event, expected_event)
-    assert_device_equal(device, expected_device)
+    utils.assert_event_equal(event, expected_event)
+    utils.assert_device_equal(device, expected_device)
 
 
 @pytest.mark.django_db
-def test_poll_provider_v0_4_archives_no_result(client, requests_mock):
-    # Note: testing with the default "POLLER_CREATE_REGISTER_EVENTS = False"
+def test_poll_provider_archives_no_result(client, requests_mock):
     provider = factories.Provider(
         base_api_url="http://provider",
         api_configuration__api_version=enums.MDS_VERSIONS.v0_4.name,
@@ -292,7 +303,7 @@ def test_poll_provider_v0_4_archives_no_result(client, requests_mock):
     )
     call_command("poll_providers", "--raise-on-error", stdout=stdout, stderr=stderr)
 
-    assert_command_success(stdout, stderr)
+    utils.assert_command_success(stdout, stderr)
     # The last poller cursor is updated
     provider = models.Provider.objects.get(pk=provider.pk)
     # No event but we won't ask this hour again
@@ -300,8 +311,7 @@ def test_poll_provider_v0_4_archives_no_result(client, requests_mock):
 
 
 @pytest.mark.django_db
-def test_poll_provider_v0_4_realtime(client, requests_mock):
-    # Note: testing with the default "POLLER_CREATE_REGISTER_EVENTS = False"
+def test_poll_provider_realtime(client, requests_mock):
     # This time we didn't poll long ago
     last_event_time_polled = timezone.now() - datetime.timedelta(3)
     provider = factories.Provider(
@@ -319,7 +329,7 @@ def test_poll_provider_v0_4_realtime(client, requests_mock):
     events = urllib.parse.urljoin(provider.base_api_url, "/events")
     requests_mock.get(
         events,
-        json=make_response(
+        json=utils.make_response(
             provider,
             expected_device,
             expected_event,
@@ -333,7 +343,7 @@ def test_poll_provider_v0_4_realtime(client, requests_mock):
     )
     call_command("poll_providers", "--raise-on-error", stdout=stdout, stderr=stderr)
 
-    assert_command_success(stdout, stderr)
+    utils.assert_command_success(stdout, stderr)
     # The last poller cursor is updated
     provider = models.Provider.objects.get(pk=provider.pk)
     assert provider.last_event_time_polled != last_event_time_polled
@@ -343,13 +353,12 @@ def test_poll_provider_v0_4_realtime(client, requests_mock):
     # The actual event
     event = device.event_records.get()
     assert event.event_type == enums.EVENT_TYPE.service_start.name
-    assert_event_equal(event, expected_event)
-    assert_device_equal(device, expected_device)
+    utils.assert_event_equal(event, expected_event)
+    utils.assert_device_equal(device, expected_device)
 
 
 @pytest.mark.django_db
-def test_poll_provider_v0_4_lag(client, requests_mock):
-    # Note: testing with the default "POLLER_CREATE_REGISTER_EVENTS = False"
+def test_poll_provider_lag(client, requests_mock):
     # Below the lag threshold
     last_event_time_polled = timezone.now() - datetime.timedelta(minutes=30)
     provider = factories.Provider(
@@ -369,7 +378,7 @@ def test_poll_provider_v0_4_lag(client, requests_mock):
     )
     call_command("poll_providers", "--raise-on-error", stdout=stdout, stderr=stderr)
 
-    assert_command_success(stdout, stderr)
+    utils.assert_command_success(stdout, stderr)
     # The last poller cursor is NOT updated
     provider = models.Provider.objects.get(pk=provider.pk)
     assert provider.last_event_time_polled == last_event_time_polled
@@ -377,85 +386,3 @@ def test_poll_provider_v0_4_lag(client, requests_mock):
     assert models.Device.objects.count() == 0
     # No event created
     assert models.EventRecord.objects.count() == 0
-
-
-#
-# Utilities
-#
-
-
-def make_response(
-    provider,
-    device,
-    event,
-    event_type_reason,
-    associated_trip=None,
-    next_page=None,
-    version="0.3.0",
-):
-    assert event.event_type in [
-        event_type for event_type, *_ in dict(PROVIDER_REASON_TO_AGENCY_EVENT).values()
-    ]
-    telemetry = event.properties["telemetry"]
-
-    response = factories.ProviderStatusChangesBody(
-        version=version,
-        data__status_changes=[
-            factories.ProviderStatusChange(
-                provider_id=str(provider.pk),
-                provider_name=provider.name,
-                device_id=str(device.pk),
-                vehicle_id=device.identification_number,
-                vehicle_type=device.category,
-                event_type=PROVIDER_EVENT_TYPE_REASON_TO_EVENT_TYPE[event_type_reason],
-                event_type_reason=event_type_reason,
-                propulsion_type=device.propulsion,
-                event_time=int(event.timestamp.timestamp() * 1000),  # In ms
-                event_location__properties__timestamp=telemetry["timestamp"],
-                event_location__geometry__coordinates=[
-                    telemetry["gps"]["lng"],
-                    telemetry["gps"]["lat"],
-                ],
-                associated_trip=associated_trip,
-                recorded=int(event.timestamp.timestamp() * 1000),  # In ms
-            )
-        ],
-        links__next=next_page,
-    )
-    assert response["data"]["status_changes"][0]["event_location"]["geometry"][
-        "coordinates"
-    ] == [telemetry["gps"]["lng"], telemetry["gps"]["lat"]]
-    return response
-
-
-def assert_command_success(stdout, stderr):
-    assert not stderr.getvalue().strip(), """Command failed!
-stdout:
-%s
-stderr:
-%s
-""" % (
-        stdout.getvalue(),
-        stderr.getvalue(),
-    )
-
-
-def assert_device_equal(device, expected_device):
-    assert device.category == expected_device.category
-    assert device.identification_number == expected_device.identification_number
-    if device.model:
-        assert device.model == expected_device.model
-    assert device.propulsion == expected_device.propulsion
-
-
-def assert_event_equal(event, expected_event):
-    # These are not (yet?) in the provider API
-    del expected_event.properties["telemetry"]["gps"]["accuracy"]
-    del expected_event.properties["telemetry"]["gps"]["altitude"]
-    del expected_event.properties["telemetry"]["gps"]["heading"]
-    del expected_event.properties["telemetry"]["gps"]["speed"]
-    assert abs(event.timestamp - expected_event.timestamp) < datetime.timedelta(
-        seconds=0.001
-    )
-    assert event.point.json == expected_event.point.json
-    assert event.properties == expected_event.properties
